// web/src/api/generated/models/PartnerVibeProjection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { PartnerVibeHidden } from './PartnerVibeHidden';
import {
    instanceOfPartnerVibeHidden,
    PartnerVibeHiddenFromJSON,
    PartnerVibeHiddenFromJSONTyped,
    PartnerVibeHiddenToJSON,
} from './PartnerVibeHidden';
import type { PartnerVibeNoCheckIn } from './PartnerVibeNoCheckIn';
import {
    instanceOfPartnerVibeNoCheckIn,
    PartnerVibeNoCheckInFromJSON,
    PartnerVibeNoCheckInFromJSONTyped,
    PartnerVibeNoCheckInToJSON,
} from './PartnerVibeNoCheckIn';
import type { PartnerVibeVisible } from './PartnerVibeVisible';
import {
    instanceOfPartnerVibeVisible,
    PartnerVibeVisibleFromJSON,
    PartnerVibeVisibleFromJSONTyped,
    PartnerVibeVisibleToJSON,
} from './PartnerVibeVisible';

/**
 * @type PartnerVibeProjection
 * Named discriminated partner Vibe state without hidden optional fields.
 * @export
 */
export type PartnerVibeProjection = { state: 'HIDDEN_UNTIL_SELF_CHECK_IN' } & PartnerVibeHidden | { state: 'NO_CHECK_IN' } & PartnerVibeNoCheckIn | { state: 'VISIBLE' } & PartnerVibeVisible;

export function PartnerVibeProjectionFromJSON(json: any): PartnerVibeProjection {
    return PartnerVibeProjectionFromJSONTyped(json, false);
}

export function PartnerVibeProjectionFromJSONTyped(json: any, ignoreDiscriminator: boolean): PartnerVibeProjection {
    if (json == null) {
        return json;
    }
    switch (json['state']) {
        case 'HIDDEN_UNTIL_SELF_CHECK_IN':
            return Object.assign({}, PartnerVibeHiddenFromJSONTyped(json, true), { state: 'HIDDEN_UNTIL_SELF_CHECK_IN' } as const);
        case 'NO_CHECK_IN':
            return Object.assign({}, PartnerVibeNoCheckInFromJSONTyped(json, true), { state: 'NO_CHECK_IN' } as const);
        case 'VISIBLE':
            return Object.assign({}, PartnerVibeVisibleFromJSONTyped(json, true), { state: 'VISIBLE' } as const);
        default:
            return json;
    }
}

export function PartnerVibeProjectionToJSON(json: any): any {
    return PartnerVibeProjectionToJSONTyped(json, false);
}

export function PartnerVibeProjectionToJSONTyped(value?: PartnerVibeProjection | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }
    switch (value['state']) {
        case 'HIDDEN_UNTIL_SELF_CHECK_IN':
            return Object.assign({}, PartnerVibeHiddenToJSON(value), { 'state': 'HIDDEN_UNTIL_SELF_CHECK_IN' } as const);
        case 'NO_CHECK_IN':
            return Object.assign({}, PartnerVibeNoCheckInToJSON(value), { 'state': 'NO_CHECK_IN' } as const);
        case 'VISIBLE':
            return Object.assign({}, PartnerVibeVisibleToJSON(value), { 'state': 'VISIBLE' } as const);
        default:
            return value;
    }
}